Fixes #12: Give signature image a `maxWidth`
The problem of the huge signature was caused by #3. Removing the fixed
`width` from the signature caused the image to scale to the page width
(even though the docs say this should mean the image's original size
*shrug*).

Instead, we now set a `maxWidth`, which solves both problems. I haven't
actually found any documentation for this feature but it was implemented
in bpampuch/pdfmake#938.